refactor: 코드리뷰 반영
- 빈 공백라인 삭제
- 메서드 뒤에 한 라인 공백 제거
- 유지보수, 가독성을 위해 컬렉션의 경우 최대한 짧게 작성하도록 수정
- 로컬 변수 삭제
- 메서드명 구분자에 적용되도록 수정
- package명 step1로 수정
