+++ src/test/java/step1/SetTest.java
@@ -1,4 +1,4 @@
-package study;
+package step1;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -28,31 +28,21 @@ class SetTest {
     @Test
     @DisplayName("HashSet의 size()는 3이 반환된다.")
     void testHashSetSize() {
-
-        // when
-        int actualSize = numbers.size();
-
-        // then
-        assertThat(actualSize).isEqualTo(3);
+        assertThat(numbers).hasSize(3);
     }
 
     @ParameterizedTest
     @ValueSource(ints = {1, 2, 3})
     @DisplayName("HashSet에 특정 값이 포함되어 있는지 확인한다.")
     void testHashSetContains(int number) {
-
-        // when & then
-        assertThat(numbers.contains(number)).isTrue();
+        assertThat(numbers).contains(number);
     }
 
     @RepeatedTest(5)
     @DisplayName("HashSet은 순서가 보장되지 않아서 실행할 때 마다 순서가 다를 수 있다.")
     void testHashSetOrderChanges() {
-
-        // when
         Integer[] array = numbers.toArray(new Integer[0]);
 
-        // then
         assertThat(array).containsExactlyInAnyOrder(1, 2, 3);
     }
 
@@ -66,11 +56,8 @@ class SetTest {
     })
     @DisplayName("HashSet의 contains()를 통해 값의 포함 여부를 반환한다.")
     void testToString(int number, boolean expected) {
-
-        // when
         boolean expectedResult = numbers.contains(number);
 
-        // then
         assertThat(expectedResult).isEqualTo(expected);
     }
 }
